fix(sip): repair pool-refusal paths that abandoned half-mutated calls
Code review of the #101(A) pass found the mechanical null-check sweep had
inserted `return` uniformly without checking what had already run. The
"drop, peer retransmits" contract only holds when nothing was mutated
before the draw; at six sites it wasn't, and the call became unrecoverable.

- onOk (hunt/broadcast answer): the draw came after setState(Connected) +
  clearRingTimer(), but the enclosing branch only re-enters while the state
  is Invited — so the 200 OK retransmit could never get back in. Caller
  never answered, losing forks never CANCELled. Draw hoisted above the
  state change.
- onMediaInvite: RTP was already streaming and the session emplaced +
  Connected before the 200 OK draw; the INVITE retransmit then hit the
  isActive() guard and got 486 Busy Here. Draw hoisted, and a refusal now
  unwinds the stream like the session-pool-full path already did.
- onBye (999/page zone): the caller's 200 OK was already queued, so it had
  its final response and would never retransmit — returning on a refused
  BYE fork skipped endCall() and leaked the session with no CDR. The fork
  is now optional; teardown always runs.
- startBroadcastFork: refusing the 180 after _sessions.emplace() left a
  session with no target invited and no ring timer to sweep it, and a
  retransmit's duplicate emplace is a no-op so it was never replaced.
  Draw hoisted above publication.
- buildInviteFork returned void, so huntRingNext still armed a 20 s
  no-answer timer and redirectInvite still reported success for an INVITE
  that was never sent. Now returns bool and callers honour it.
- redirectInvite returned false on refusal, which means "target not
  registered, fall through" — on the CFU path that rings the extension the
  subscriber forwarded away from. Now returns true, matching its sibling
  503 branch.

Also from the review:
- RegisterBeeper no longer advances to AwaitingByeOk when the BYE was
  refused; sweep() would have freed the slot having never sent one,
  leaving the phone off-hook in a forgotten call.
- BlfSubscriptions::sweepExpired keeps the slot when the terminal NOTIFY
  is refused, instead of wiping it and freezing the watcher's lamp.
- tick() stamps setLastPingTime only when a ping was actually built.
- Restored the early-warning log lost in the first pass: "falling back to
  heap" and "dropping" are distinct signals again, with a running total,
  and both overloads now share one rate-limit counter per pool.
- _bodyGen widened to uint64_t so it cannot wrap onto a generation a cache
  already recorded, or onto the 0 sentinel.
- Acquire fence after observing use_count()==1: the mutex serializes
  takers, but the previous owner released outside it and reset() reads the
  message's existing internals before overwriting them.

188/188 host tests; all 7 edited files cross-compile clean for ESP32-S3.

+++ src/SIP/SipMessage.hpp
@@ -139,7 +139,7 @@ protected:
 	//
 	// Not atomic, and not internally synchronized: like the rest of SipMessage,
 	// a given message is only ever touched under RequestsHandler::_mutex.
-	uint32_t bodyGeneration() const { return _bodyGen; }
+	uint64_t bodyGeneration() const { return _bodyGen; }
 
 	bool _hasSdp = false;
 
@@ -176,7 +176,13 @@ private:
 	// The copy CONSTRUCTOR does copy it, which is correct — there is no
 	// pre-existing cache in a brand-new object to go stale, and the copied body
 	// and copied cache describe the same bytes.
-	uint32_t                 _bodyGen = 1;
+	// 64-bit, not 32: it only ever increments, and a pooled slot on a
+	// long-lived device bumps it several times per handled message. A 32-bit
+	// counter would eventually wrap onto a value a derived cache had already
+	// recorded — reviving a stale SDP parse against a different call's body, the
+	// exact hazard this exists to prevent — or onto 0, the fresh-cache sentinel.
+	// At even 10k bumps/second a 64-bit counter outlives the hardware.
+	uint64_t                 _bodyGen = 1;
 
 	sockaddr_in _src{};
